Fixed unhandled exception on missing parameters
It seems that the SUPINFO OpenID provider no longer returns most of the profile information (except the BoosterID and the full name) which causes the module to throw an error when it attempts to split `PersonCampus`. This pull request will prevent the module from crashing the app and will instead set `profile.campus` and `profile.campusId` to `'N/A'`.